services/mt5-bridge/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
python-dotenv==1.0.1
MetaTrader5==5.0.45; sys_platform == "win32"
